% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotsign-raw.R
\name{shotsign_raw}
\alias{shotsign_raw}
\title{Exploratory visualization for 3 Quantitative Variables}
\usage{
shotsign_raw(data, x, y, color, ...)
}
\arguments{
\item{data}{\code{data.frame} with data to plot}

\item{x}{the variable to plot on the x axis.  This should be the unquoted 
name of the column in \code{data} containing the x values.}

\item{y}{the variable to plot on the y axis.  This should be the unquoted 
name of the column in \code{data} containing the y values.}

\item{color}{the variable to use to color the plot.  This should be the 
unquoted name of the column in \code{data} containing the color values.}

\item{...}{additional arguments passed to \code{\link{shotsign}}.}
}
\description{
Exploratory visualization for 3 Quantitative Variables
}
\details{
Estimates the density of \code{x} using \code{\link{density}}. 
  \code{density(x)$x} will be used as the x values in \code{\link{shotsign}},
  and \code{density(x)$y} will be used as the width values.
  
  Fits Generalized Additive Models (GAMs) for \code{y} and \code{color} using
  \code{\link[mgcv]{gam}}.  These models are specified as \code{gam(y ~ s(x,
  bs = "ts", k = 7))} and the predictions are used to generate the y and
  color values for \code{shotsign}.
}
\examples{
shotsign_raw(iris, x = Sepal.Length, y = Petal.Length, color = Sepal.Width)
}
\seealso{
shotsign
}

